renamed 'seeders' folder to 'seeds'
