+++ prisma/seed.ts
@@ -1,6 +1,6 @@
 import { sectors } from "./seeds/sectors";
 import {prisma} from "../lib/prisma";
-import AdminSeed from "./seeders/admin";
+import AdminSeed from "./seeds/admin";
 
 async function main() {
   await prisma.sector.createMany({
